server/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
python-multipart==0.0.6
aiofiles==23.1.0
openpyxl==3.1.2
flask
flask-cors
flask-sqlalchemy
